refactor(coverage): relocate 4 test-helpers under testing/ (self-enforcing exemption)
Drops their explicit exclusions; discoverSourceFiles auto-skips /testing/.
